+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -11,10 +11,10 @@ import org.springframework.http.MediaType;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationSaveRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.dto.ThemeResponse;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.request.ReservationSaveRequest;
+import roomescape.dto.response.ReservationTimeResponse;
+import roomescape.dto.response.ThemeResponse;
 import roomescape.exception.NotFoundException;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
